Refactor how Regs and TransientRegs share attributes
Initially I had TransientRegistration just inherit from Registration. However, this put them in the same table, which was not desirable.

This tweak allows both types of Registration to share most attributes and validations but still have some unique attributes if required. And TransientRegistrations get their own collection in the database.