Set up 'happy path' for limited company renewals
https://eaflood.atlassian.net/browse/WC-124

The 'happy path' should cover the pages which the user will need to complete a renewal, provided they meet certain conditions (in this case, a limited company based in the UK).

This will be in a fairly skeletal form - just a heading, a back link and a continue button. It should provide a basis for testing and further development. Functionality like the address lookup and smart answers is not covered.

This will also handle implementing a state machine to manage how the user travels between pages.