Rename states to avoid clash with fields
Several of the workflow states had the same name as a field name (eg contact_name). This was throwing up warnings from Rails as it caused some method clashes (eg contact_name?). I've added _page to all states to avoid a clash.

Also tweaked a test as we now have a neater way of creating valid TransientRegistrations for tests.