Move to next workflow state when forms are submitted
This is still in a bit of a state (ha, ha) but want to commit before it gets any bigger. When a form is submitted, if the contents are valid, the "next" event should fire and they should be redirected to the correct next form. Form models and controllers should be inheriting behaviour from a base model and controller.

Also renamed the states to match the form names.

Users going to forms in the "wrong" order should also be redirected to the state which the renewal is currently in. This doesn't really work on POST yet.