Set up form for business type
Add form options for selecting the business/organisation type. If what's in the DB matches an option, this will be pre-selected.

One test is failing – this is because I set the TransientObject's after_initialize method to only happen when the record is new. Will fix this when I'm back from my hols 🎄