Check if change to business types are valid
https://eaflood.atlassian.net/browse/WC-91

Users have the option to change their business/org type for the registration. However, some of these changes will result in the user needing to make a new registration, rather than renewing the existing one.

If this is the case, we should explain this to the user and direct them to the new registration process. Otherwise, they can continue with the renewal.