Set up "skeleton" of happy path for local authority/public body
https://eaflood.atlassian.net/browse/WC-146

This story covers setting up a "skeleton" of the happy path for those who would select Local authority or Public body on the select org type page. Pages in the skeleton will mostly be blank – just a title and continue button.

This is so we can set up how users move between the pages, and also so we can start testing the process as a whole.

As we continue to work on the stories, these pages will be fleshed out with the proper forms.