Improve database seeding
https://eaflood.atlassian.net/browse/WC-167

We need some seed data for the database for the dev environment and for running acceptance tests against. This change switches from loading all seeds in a single script to storing seeds as multiple JSON files and then loading them. This approach should make it easier to add and modify seeds.