Add individual forms for smart answers
These forms were generated using the scaffolding tool. Also removed the "smart answers" holding page which was filling in for the different forms. Some tests are still failing as they are relying on data to be submitted through the forms - will sort these out shortly but wanted to commit before the PR got too big.