Bring validation errors up to GOV.UK standards
Validation errors should now meet the standards in http://govuk-elements.herokuapp.com/errors/

Elements should now display correctly, and the errors at the top of the page link to the specific fields with the errors.

This included a fix for the business type form as well as this was the other implemented one.