Refactor business_type_form state tests
These were especially repetitive as we're just feeding in different values and making sure we get the correct result. I switched to using a permutation table pattern as described in this talk: https://www.youtube.com/watch?v=KjENZNjRCWM This allows us to just store the values and the expected result in a single hash, and then run them all through the same process.