Get the registration number of a limited company or LLP
https://eaflood.atlassian.net/browse/WC-97

Limited companies and LLPs must provide their registration number for Companies House. We then need to validate that the number is in the correct format, then query Companies House to make sure the company/LLP is active. We should then inform the user if Companies House says there's a problem, or return the details from Companies House so the user can confirm all is well.