Provide an address manually
https://eaflood.atlassian.net/browse/WC-101

While we prefer users to add addresses using the OS Places API, some users will not be able to use it, for example if they are overseas, if their address is not listed or if the service is down. These users should be able to fill in a separate form with details of their address.