Only prefill address form if we don't want to use temp_postcode
If the user provides an address, saves and then goes back to the form later, we want to fill in this address.

However, if the user has since changed the postcode on a different form, we don't want to fill in the old address as this will have changed. Instead we should just fill in the postcode which they provided.

This does not apply to overseas users, who would never be setting a postcode on a different form.