Users can enter address after postcode search returns results
If the OS Places lookup returns addresses but the user doesn't want to use any of them, they can choose to go to the manual address form instead.

Once they have saved an address, the form will pre-fill with the address if they return to the form, UNLESS they have subsequently changed the postcode on a previous page (thus invalidating the saved address).