Rework OS places error routing, add flash message
After some discussion, decided to rip out the transition_flag we were using for routing and just temporarily persist a flag for the error against the transient_object. This flag is then removed once the manual address form is loaded. Not ideal, but also means less mucking around in the controller, and the way transitions are handled is much more consistent.