Tweak validation handling and heading
These changes are following the QA review. The bulk of the changes are about how we handle validation messages. This commit reduces the number of validation errors which a user can get for submitting the form without all the required information. Previously a totally blank form would trigger about 6 different messages - now it should just be one. Date validation errors have also been simplified so if all fields are missing, we only display a single message. We also only check that the complete date is a valid date once we've validated the individual fields to reduce the number of errors users get.

Also changed the heading text for overseas users.