Validate minimum number of people
Partnerships must have at least 2 key people. All others must have at least 1. This change adds validation to make sure enough people have been added before proceeding to the next stage.

If the user attempts to add a completely blank person, we check to see if the minimum number of people has already been reached. If it is, the form validates but does not add the blank person. If it does not validate, the user is prompted to add another person.

If the user must add more people, the "Continue" button isn't visible yet.

Also reworked how the date fields are processed to make sure we set nil values if the date fields aren't in the correct format - otherwise it didn't overwrite any existing values (eg if a sole trader saved a key person, then tried to edit it).