Improve button styling
Style the 'Add another' link as appropriate for the form. Sometimes this will be the main action (eg when adding the first partner) but most of the time it's a secondary action.

Also hide the "You have added the following people" heading unless the business type allows multiple key people and at least one person has been added.