Check which country users are based in
https://eaflood.atlassian.net/browse/WC-238

This PR adds a new form to ask where the business is based - England, Wales, Scotland, Northern Ireland or overseas.

Wales, Scotland and Northern Ireland have their own registers for waste carriers. This form will allow us to direct those users to the correct service which they should be registering with. Users can continue to register with the EA if they want, but this will at least make them aware of the other registry.

Overseas users operating in England should register with the EA - however, there are some differences in what forms they see (for example, the address forms). Currently we filter them out using the 'overseas' option on the business type form, but 'overseas' makes a lot more sense as a location than a business type.