Remove checks for 'overseas' business type
Since we are now flagging overseas users based on the location, we should check that field instead of business_type to filter for overseas users. This commit changes a lot of logic which was relying on checking the business_type and instead introduces an `overseas?` method which relies on the location instead.

We do continue to set the business_type to overseas as this is used for microcopy in our app, and is expected to be filled in when other applications access registrations from the shared database. However, we now set this in the location form if the location is overseas.