Declare if a relevant environmental offence exists
https://eaflood.atlassian.net/browse/WC-107

"Has anyone involved in managing your organisation been convicted of an environmental offence in the past 12 months?" Users can only answer yes/no on this form – if they say yes, they should be sent to an additional form to provide further details. (That form is a separate story.)

We should not prepopulate this form with their previous answer, as this only applies to the last 12 months, so an answer from 3 years ago is no longer relevant.