Users can add people with relevant convictions
https://eaflood.atlassian.net/browse/WC-108

If a user answers yes to "Has anyone involved in managing your organisation been convicted of an environmental offence in the past 12 months?" they should be prompted to provide the details of the person with the offence (but not the details of the offence).

This should be a blank form as any data from the previous registration is out of date.