Distinguish between 'key' and 'relevant' KeyPeople
The KeyPeople class actually covers two types of people:

- people with the 'key' person_type, added on the key people form
- people with the 'relevant' person_type, added on the relevant conviction details form

For what I assume are historical reasons in the existing database, all of a registration's people are stored together as KeyPeople embedded objects, with the only distinguishing factor being the 'person_type' attribute. However, as far as the frontend is concerned, these are two separate lists. We need to make sure we don't modify key people when changing relevant people, and the reverse. For example, checking the number of key people which a registration has (eg, a sole trader can only have one) should not also count 'relevant' people.

This change adds some methods to registration-y objects to get lists of 'key people' or 'relevant people' only. It also makes some modifications to the existing key people form to make sure we leave relevant people alone when adding, listing or deleting key people.

And, of course, we really needed some tests for all this.