Set up form for conviction details
The forms for key people and conviction details are very similar - the only difference is the additional "position" field for the conviction details form. So to start out with, we're mostly duplicating key people and making sure the tests pass.

Once we have the form and test suite in place, we can start reducing duplication between the two (and improving some of the naming).