Clean data when creating and saving transient registrations
This change adds some features to improve our data quality.

This means removing excess whitespace at the start or end of the field. Some forms do this on saving but others don't, so we should try to implement this across all forms in a standardised way instead. We should also do this when we import data from a Registration to a TransientRegistration object.

There are also some fields, like phone numbers, which previously have not been validated as closely as we would have liked. It doesn't make sense to prefill a field with a value which will then be considered invalid, so we should run some validations on importing the data when the TransientRegistration is created.