Remove invalid attributes from brand new transient registrations
When we create a new transient registration, we import attributes from the source registration. Due to lax validations in the past, there's a possibility that some of this data may not pass our new validations. Prefilling a field with data which does not validate is frustrating for the user so we've chosen to remove it instead. For more info on this decision, see https://eaflood.atlassian.net/browse/WC-131

Phone numbers were the main reason for this, so I've also implemented phone_number validation so we have something to test with. I used the Phonelib gem for consistency with our other Rails apps (WEX and FRAE).