Validate email fields
First we check that the fields are filled in. Then we check if the email format is correct, and if the two fields match.

We only check the format of the contact_email field, not the confirmation field, as that only really exists to make sure they match. If contact_email is a correct format and confirmed_email is identical to that, then it would also be a correct format.

The email validation gem also covers length, so we don't need an additional check for that.