Add application version and current commit
This will need further work when we move the code into an engine - we'll need to keep track of the engine version separately from the frontend and backend versions. But for now it's simple enough to just pop in the config.

The current_git_commit code is cribbed from waste-exemptions-shared.