Checks for can_renew include expires_on
https://eaflood.atlassian.net/browse/WC-185

Currently there will be a brief window when a registration is officially EXPIRED but it is still recorded as ACTIVE in the service. This is due to the timing of when the job that marks expired registrations is run.

To avoid any issues with someone renewing before the expiry job runs, we should make sure that a registration cannot be renewed if the current date is equal to or after the expires_on date.