Account for daylight savings time when checking if expired
When we check if a registration should be expired, we want to compare against the current date in the UK. This may be different than the current date in UTC when daylight savings is in effect.

If a user registers around the time of year when DST usually starts or ends, there's also a chance that they may be registering in one time zone and expiring in another. So we need to account for this to make sure registrations don't gain or lose a day.