Make smart answers optional
https://eaflood.atlassian.net/browse/WC-275

In order to reduce the number of pages users have to go through, and to give more choice for those who have good knowledge of what the difference is between lower and upper tier, we will make smart answers optional.

We have already tested this in user research without issue, and the team feels it better highlights that smart answers is an aid and not an approval or clarification of what users can do (which testing has shown some view it as).

This story covers implementing the version contained in the prototype into the production renewals service.