Correct mongoid.yml config file format
Since we first built the renewals service we have been unable to connect to MongoDb with authentication switched on.

Having investigated the issue, it all appears to be due to changes in the format of the `mongoid.yml` file caused by Mongoid being taken in house (by MongoDB), and replacing the underlying driver Moped with the official ruby MongoDb one.

This changes resolves those formatting issues, and tests confirm that it can now connect to a MongoDb 3.6 database with authentication enabled.

WIP