Persist and validate number of cards to order
We have decided to hold off on implementing WorldPay for now and focus on getting as much of the journey done as possible for user testing first. So for now, we will just save the number of cards as a temporary attribute.

The field is populated with 0 by default but can be changed to any positive integer (or left blank).