Add payment summary page
https://eaflood.atlassian.net/browse/WC-116

Before they are asked for payment, users should see a summary of the amount they must pay and their options for paying it.

This should include the renewal fee, the cost of any copy cards, and the fee for changing the registration type (if applied).

This does not cover setting up the order or payment objects in the database as this will be dealt with when WorldPay is implemented.