Pay for a renewal by bank transfer
https://eaflood.atlassian.net/browse/WC-118

If a user chooses to pay by bank transfer on the Payment Summary page, they should be sent to an additional page which has all the necessary instructions, information and bank details for them to make the payment.

This is instead of going through WorldPay.

It should be made clear that paying by bank transfer means you cannot instantly renew your registration as you have to wait for your renewal to be processed.