Complete a renewal with a pending conviction check or payment
https://eaflood.atlassian.net/browse/WC-120

If a user submits their renewal and pays by bank transfer, or submits and has a conviction check pending, they should see a "Renewal received" page.

This story covers setting up that page and adding routing to it depending on the payment type. It does not cover the conviction checks.