Reroute users with declared convictions
If a user declares they have convictions, we need to do some extra checks before renewing their registration. So they should be routed to the renewal_received_form instead. In the future this will cover any registrations flagged by the entity matching service as well.

Because a user could end up on the page because of their payment method, having a pending conviction check, or both, we need to make sure they only see content relevant to them.