Make sure can_renew checks are actually applied
Our state machine for registrations has several checks to see if a registration can go through the renew action. However, we don't actually have this implemented anywhere (oops). This change makes sure that we are actually putting those checks into practice, so if a user bypasses the registration search in the frontend and accesses the renewal forms directly, they still can't renew something which should not be renewable.