Make the app play nice with the browser back button
https://eaflood.atlassian.net/browse/WC-307

Viewing or submitting forms is currently only allowed if the saved workflow_state matches the form in question.

This should be fine if the user is only navigating by the buttons in the app, but using the browser back button to access earlier, cached forms means the user can be looking at a different form than what the app thinks they should be looking at.

Submitting one of these cached forms will send the uesr back to the form matching the workflow_state, but won't save anything they submit, or warn them about that.

We should modify how the state machine works to make sure users can still navigate with the back button and save their changes.