Start shared_examples_for ideal form request behaviour
These new shared examples describe the behaviour we are aiming for with the changes in this branch. We haven't written or implemented any tests yet, but this sets out the plan for what should happen.

This should also reduce some duplication in our request tests, as there is a lot at the moment.