Start writing tests for form requests, split files
Add shared example tests for viewing (GET) and submitting (POST) forms. There are different sets of GET tests depending on the desired behaviour of the forms. A 'locked-in' form means that we expect the form the user is accessing to match the workflow_state in the transient_registration, and if not, we will redirect to the workflow_state. This should be how all forms behave currently, so we should be able to swap in the new tests and still have everything pass.

As we change the behaviour of the forms, we will bring in the new tests.

I also split the tests into separate files to make things more readable.

These were tested against the location form. However I will swap the existing tests for the shared examples in a separate commit, as that will involve a LOT of file changes.