Implement flexible form behaviour
When a user makes a GET request to a form which doesn't match the saved workflow_state, and both the requested state and the workflow_state are considered to be 'flexible forms', load the requested form instead of redirecting to the workflow_state.

Otherwise, we redirect to the form for the saved workflow_state as we did before.

Both the saved and requested states must be flexible - otherwise we risk issues with users changing states when they shouldn't, for example from the renewal_complete form, or skipping ahead to it.