Flag form objects as flexible states instead of listing them
Previously we stored a list of workflow_states considered to be flexible.

After some discussion, we decided to try marking the form objects which match the workflow_state as flexible instead. This is done by including the new CanNavigateFlexibly concern. Then we check for the presence of the concern instead of checking if the state is included in the list.