Only validate company_no if it's required
Now that we're validating company_no on the 'check your details' page as well as the form where it's entered, we could be trying to validate it for business types which don't actually need it. This tweaks the validations to check the business type first and then validate as required.