Validate people on 'check your details'
We need a separate validator for checking people on the 'check your details' page compared to the forms - on the forms, we just validate the data submitted for a person one at a time, whereas in this case we need to check all the people (and make sure the numbers of people are correct). This meant quite a lot of refactoring, including moving all methods surrounding the number of people per type to a separate concern.