Validate addresses on 'check your details' form
Validate that addresses exist and also are in the correct mode for the location (so if you add a UK address, then go back and change your location to overseas, your UK address is no longer valid).

We don't use the new validator on the address forms as these have separate validations on the content being submitted.