Validate MAC when we get a success or failure from WorldPay
We store a MAC on WorldPay and in our .env file. We can use this to verify that the requests we get when a user is redirected back from WorldPay are valid (rather than someone attempting to fake a payment by URL-hacking).

This code was mostly cribbed from the frontend app. However, for a similar implementation in a more up-to-date verson of WorldPay, see:
http://support.worldpay.com/support/kb/gg/corporate-gateway-guide/content/hostedintegration/securingpayments.htm