Update registration after completed renewal
https://eaflood.atlassian.net/browse/WC-303

The user has entered all their details, paid their fees and not triggered any conviction checks. The renewal has reached the "Renewal complete" stage of the forms. Now it's time to actually renew the existing registration.

The end result should be that the registration is updated to include all information added or confirmed during the renewals process, with the previous version of the registration stored as a nested object. The exception is finance_details, which should always include all past orders and payments as well as the latest ones.