Start implementing RenewalCompletionService
This service will handle all the steps to update a registration when renewal is completed. Also add some tests to set out what the end result should be.

We have some issues with copying data around the registration - for example, from the registration to a nested past_registration, or a transient_registration to a registration. This is currently being fudged but will need further work.