Persist past_registration when creating it
Trying to persist the past_registration in the service was causing issues, so save it on creation instead. But first we check to make sure a past_registration for this version of the registration doesn't already exists (to avoid duplication if we hit an error later in the process and have to start over).