Replace values of unanswered questions with nil
If the user skips a question (for example, smart answers), the transient registration won't set a value for that attribute. But if they had an earlier answer which is saved to the registration, we want to overwrite that with a nil. Otherwise we risk old, irrelevant answers getting mixed in with the new data.