Create an order in the database when paying by bank transfer
https://eaflood.atlassian.net/browse/WC-395

If a user pays by WorldPay, we create an order in the database before we create the corresponding payment.

However, bank transfers don't log any record in the database at the moment.

Choosing to pay by bank transfer should create a new order in the database containing all the relevant information.

If there is already an order created by someone attempting to pay by WorldPay (but not succeeding), this order should be overwritten by the new one so we don't get any duplication.