Specify payment method when initialising FinanceDetails
When we initialise a new FinanceDetails, we now specify whether it's for a WorldPay or bank transfer payment, and set up the nested order accordingly.

This means we can now create a new FinanceDetails when the user hits the bank transfer page. If there is any FinanceDetails info hanging around already, for example if the user attempted ot pay by WorldPay first and then cancelled, this is overwritten.