Set up EntityMatchingService
Create a new service to send entity matching requests to the Waste Carriers Service API. So far this just formats the correct URLs to send the query. For now, we return the URLs so we have something to test - this is just temporary.

This commit also sets up some extra date formatting, env vars and test factories to prepare for the request.