Save entity match result to transient_registration
If the result is for the business, it's stored on the transient_registration itself. If it's for a person, it's stored under that person.

We save the result whether it's a positive or negative match. If we run into an error when sending our query to the Java service, we save an "unknown" result.