Set default values for the various charges
Previously in the frontend the charges were hard coded. Now like renewals they come from the same environment variables. However unlike the frontend renewals defaulted the values to 0 if the env vars were not set.

This change brings renewals inline with the frontend, by defaulting the charge values to what they actually are (and have been since the service launched).