Move /app files to engine
Move controllers, forms, helpers, models, services and views which should be part of the engine to the folder. We don't make any changes to them yet as we want Git to recognise that the files have moved, not been deleted and recreated - that way we can hopefully keep the history.

Will this break the build? 100%.