Move required classes and modules to WasteCarriersEngine
Move controllers, forms, models, helpers, services and validators to the WasteCarriersEngine.

While we're in here, add `# frozen_string_literal: true` to appease Rubocop. We also remove some stray comments and .keep files which we no longer need.

The ability class also should not be part of the engine as it relates to the users, so that is moved back to the correct folder.