Require AASM in gemfile
In WasteCarriersEngine::CanChangeWorkflowStatus, we need to `include AASM`. This worked fine before it was moved into an engine but doesn't seem to work properly now unless we include AASM in the Gemfile as well as the gemspec.

To unblock fixing other issues, we'll do this for now - however I think there should be a better solution which will be looked into shortly.