Move User and Ability into WasteCarriersEngine
Move the functionality from Devise and CanCanCan into the engine. Devise seems to be working now. CanCanCan is not, so the tests are a nightmare again, but committing to save the progress so far.

It also turns out there's an easier way to add Devise helpers to request specs, so we can ditch RequestSpecHelper entirely.